Moving Graph calls closer to calling component
